requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
distro==1.9.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
numpy==2.2.5
openai==1.6.0
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
